# packages/backend/openapi/openapi.yaml
...
openapi: '3.1.0'
info:
  title: RONL Business API
  # NO `version` here, deliberately. scripts/build-openapi.cjs sets it from
  # package.json and REFUSES TO BUILD if this file sets one, so the published
  # document and the release can never disagree.
  summary: Process, decision and document services for Regels.overheid.nl
  contact:
    name: Steven Gort
    email: steven.gort@ictu.nl
    url: https://iou-architectuur.open-regels.nl/
  description: |
    Business API behind Regels.overheid.nl: Operaton process and decision
    execution, task handling, document generation and signing, and the
    machine-to-machine surface consumed outside the browser applications.

    ## Authentication

    Most operations require a Keycloak-issued JWT as a bearer token, applied by
    `jwtMiddleware` inside each router rather than at the mount. A tenant is
    derived from the token and enforced per operation; some operations
    additionally require a role.

    The unauthenticated surface is deliberate and small: `/v1/health`,
    `/v1/openapi.json`, `/v1/public`, `/v1/media-aggregator`, and the ValidSign
    callback, which carries no token because ValidSign does not send one.

    ## Error responses

    Every 4xx and 5xx is RFC 9457 problem details, served as
    `application/problem+json` (#216): `type`, `status`, `title`, `detail` and
    `instance`, plus a `code` extension member -- the stable identifier to
    branch on (see the `Problem` schema). Until #216 errors answered
...

...

...

...
servers:
...
paths:
...
  /process/{key}/start:
    post:
      operationId: startProcess
      tags: [Process]
      summary: Start a process instance
      description: |
        Starts the process deployed under the caller's own organisation.

        **Which organisation gets the case** depends on who is asking:

        - **Staff** may only start a process deployed under their own
          organisation. Starting another's answers `403 TENANT_MISMATCH`.
        - **Citizens** are not refused. Their case is assigned to the
          organisation the process is deployed under, and `originTenantId`
          records their own.

        Every instance is stamped with two variables that cannot be changed
        afterwards: `municipality` (the owning organisation, which every access
        check enforces) and `originTenantId` (the caller's organisation at
        start).

        A minted `businessKey` is prefixed with the OWNING organisation, so
        the case's human-facing handle names whoever is handling it (#234). A
        caller-supplied `businessKey` is kept exactly as given — a RIP phase
        inherits the key its R2.1 run minted, so every phase of one project
        shares it. Either way the key grants nothing: access runs on
        `municipality` alone.
      parameters:
        - $ref: '#/components/parameters/ProcessKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                businessKey:
                  type: string
                  description: |
                    Optional. When absent the service mints one, prefixed with the
                    OWNING organisation. When supplied it is kept verbatim, which is
                    how a RIP phase inherits the key its R2.1 run minted (#234).
                variables:
                  $ref: '#/components/schemas/PlainVariableMap'
      responses:
        '201':
          description: The instance was started.
          content:
            application/json:
              schema:
                type: object
                required: [success, data]
                properties:
                  success: { type: boolean, const: true }
                  data: { $ref: '#/components/schemas/ProcessStartResult' }
          headers:
            API-Version:
              $ref: '#/components/headers/ApiVersion'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/TenantMismatch'
        '409':
          $ref: '#/components/responses/AmbiguousDeployment'
        '500':
          description: |
            Operaton refused the start (`PROCESS_START_FAILED`). Two extension
            members say why: `details`, Operaton's own message, and `engine`,
            the base URL of the engine that refused. (`engine` was
            `error.instance` before #216; `instance` is now the request path.)
          $ref: '#/components/responses/Error'
  /process/history:
    get:
      operationId: getProcessHistory
      tags: [Process]
      summary: Completed instances for one applicant
      description: |
        Requires `applicantId` — there is no unfiltered history.

        Restricted by ROLE, not by organisation, so it answers `403 FORBIDDEN`
        rather than `TENANT_MISMATCH`:

        - a citizen may request only their own history;
        - a caller who is neither a citizen nor a caseworker is refused
          outright.
      parameters:
        - name: applicantId
          in: query
          required: true
          description: The applicant whose history to return.
          schema: { type: string }
      responses:
        '200':
          description: The applicant's completed instances.
          content:
            application/json:
              schema:
                type: object
                required: [success, data]
                properties:
                  success: { type: boolean, const: true }
                  data:
                    type: array
                    items:
                      type: object
                      description: A historic process instance as Operaton reports it.
                      additionalProperties: true
          headers:
            API-Version:
              $ref: '#/components/headers/ApiVersion'
        '400':
          description: '`applicantId` was not supplied (`VALIDATION_ERROR`).'
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: |
            A role refusal (`FORBIDDEN`): a citizen asking for another
            applicant's history, or a caller who is neither citizen nor
            caseworker.
          $ref: '#/components/responses/Error'
        '500':
          $ref: '#/components/responses/Error'
  /process/{id}/status:
    get:
      operationId: getProcessStatus
      tags: [Process]
      summary: Whether an instance is running
      description: Admits the owning organisation and the case's own applicant.
      parameters:
        - $ref: '#/components/parameters/ProcessInstanceId'
      responses:
        '200':
          description: The instance.
          content:
            application/json:
              schema:
                type: object
                required: [success, data]
                properties:
                  success: { type: boolean, const: true }
                  data: { $ref: '#/components/schemas/ProcessStatus' }
          headers:
            API-Version:
              $ref: '#/components/headers/ApiVersion'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/TenantMismatch'
        '404':
          description: No such instance (`PROCESS_NOT_FOUND`).
          $ref: '#/components/responses/Error'
  /process/{id}/variables:
    get:
      operationId: getProcessVariables
      tags: [Process]
      summary: Current variables of a running instance
      description: |
        Admits the owning organisation and the case's own applicant.

        Values are returned PLAIN — Operaton's `{value, type}` wrapper is
        unwrapped. Requests that SET variables use the wrapped form; see
        `OperatonVariableMap`.
      parameters:
        - $ref: '#/components/parameters/ProcessInstanceId'
      responses:
        '200':
          description: The variables.
          content:
            application/json:
              schema:
                type: object
                required: [success, data]
                properties:
                  success: { type: boolean, const: true }
                  data: { $ref: '#/components/schemas/PlainVariableMap' }
          headers:
            API-Version:
              $ref: '#/components/headers/ApiVersion'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/TenantMismatch'
        '404':
          description: No such instance (`PROCESS_NOT_FOUND`).
          $ref: '#/components/responses/Error'
  /process/{id}/historic-variables:
    get:
      operationId: getProcessHistoricVariables
      tags: [Process]
      summary: Variables of a finished instance
      description: |
        The same shape as the live read, for an instance that has ended.
        Admits the owning organisation and the case's own applicant.
      parameters:
        - $ref: '#/components/parameters/ProcessInstanceId'
      responses:
        '200':
          description: The variables.
          content:
            application/json:
              schema:
                type: object
                required: [success, data]
                properties:
                  success: { type: boolean, const: true }
                  data: { $ref: '#/components/schemas/PlainVariableMap' }
          headers:
            API-Version:
              $ref: '#/components/headers/ApiVersion'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/TenantMismatch'
        '500':
          description: '`HISTORIC_VARIABLES_FAILED`.'
          $ref: '#/components/responses/Error'
...
  /process/{instanceId}/decision-document:
    get:
      operationId: getProcessDecisionDocument
      tags: [Process]
      summary: The decision document generated for a case
      description: |
        Admits the owning organisation and the case's own applicant. Most
        processes have no document template, in which case this answers 404
        rather than an empty document.

        **This operation does not use the `{ success, data }` envelope.** It
        answers `{ success, template }` -- `process.routes.ts` sends
        `res.json({ success: true, template })`. Published as `data` until
        #214, which is the kind of error only reading the handler finds: no
        test ever named the field.

        The value is the template WITH ITS BINDINGS, not a rendered document.
        Nothing here substitutes the placeholders.
      parameters:
        - name: instanceId
          in: path
          required: true
          description: The process instance id.
          schema: { type: string }
      responses:
        '200':
          description: The document template.
          content:
            application/json:
              schema:
                type: object
                required: [success, template]
                properties:
                  success: { type: boolean, const: true }
                  template: { $ref: '#/components/schemas/DocumentTemplate' }
          headers:
            API-Version:
              $ref: '#/components/headers/ApiVersion'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/TenantMismatch'
        '404':
          description: |
            No document template for this process (`DOCUMENT_NOT_FOUND`). The
            usual answer for a process that generates none.
          $ref: '#/components/responses/Error'
        '500':
          description: '`DOCUMENT_FETCH_FAILED`.'
          $ref: '#/components/responses/Error'
  /process/{key}/start-form:
    get:
      operationId: getProcessStartForm
      tags: [Process]
      summary: The deployed start form for a process
      description: |
        Returns the form-js schema a caller renders to start this process.

        No organisation check: a form is a template, not a case. It can still
        answer `409` when the key is deployed under several organisations and
        none is the caller's, because there is then no form to choose.
      parameters:
        - $ref: '#/components/parameters/ProcessKey'
      responses:
        '200':
          description: The form schema.
          content:
            application/json:
              schema:
                type: object
                required: [success, data]
                properties:
                  success: { type: boolean, const: true }
                  data: { $ref: '#/components/schemas/FormSchema' }
          headers:
            API-Version:
              $ref: '#/components/headers/ApiVersion'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: The process has no deployed start form (`FORM_NOT_FOUND`).
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/AmbiguousDeployment'
        '415':
          description: |
            The deployed form is not a form-js schema — an embedded or external
            form this service cannot serve (`UNSUPPORTED_FORM_TYPE`).
          $ref: '#/components/responses/Error'
        '500':
          description: '`FORM_FETCH_FAILED`.'
          $ref: '#/components/responses/Error'
  /process/{key}/variable-hints:
    get:
      operationId: getProcessVariableHints
      tags: [Process]
      summary: Variable names and types a process expects
      description: |
        **This operation does not use the `{ success, data }` envelope.** It
        answers `{ success, variables }`. Recorded as it is rather than
        normalised, because a consumer reads what the handler sends.
      parameters:
        - $ref: '#/components/parameters/ProcessKey'
      responses:
        '200':
          description: The hints.
          content:
            application/json:
              schema:
                type: object
                required: [success, variables]
                properties:
                  success: { type: boolean, const: true }
                  variables:
                    type: array
                    items:
                      type: object
                      required: [name, type]
                      properties:
                        name: { type: string }
                        type: { type: string }
          headers:
            API-Version:
              $ref: '#/components/headers/ApiVersion'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          description: '`VARIABLE_HINTS_FAILED`.'
          $ref: '#/components/responses/Error'
...
components:
...
  securitySchemes:
    m2mOAuth:
      type: oauth2
      description: |
        The machine-to-machine surface, `/v1/m2m`.

        The token is obtained with a CLIENT CREDENTIALS grant against the same
        realm the rest of the API uses, and is validated the same way: same
        issuer, `ronl-business-api` audience, RS256. It is a separate scheme
        here because the CALLER is checked as well as the token.

        **The client must be allow-listed.** `requireM2mClient` requires the
        token's `azp` -- the party it was issued to -- to appear on the server's
        `M2M_ALLOWED_CLIENTS`, which is `operaton-mcp-client` alone today.
        Anything else is refused with `403 M2M_CLIENT_NOT_ALLOWED` before any
        handler runs. Adding a consumer means adding its client id to that
        variable; nothing changes in Keycloak.

        `azp` rather than the audience, because every person in the realm holds
        a token for `ronl-business-api`: the audience cannot distinguish a
        machine from a citizen, and before #237 that meant any authenticated
        user reached all 18 operations.

        OpenAPI cannot express the allow-list condition, so it is stated here
        and carried by the 403 on every operation. A bearer token that is
        merely valid is NOT sufficient.

        `tokenUrl` below is production's. Acceptance issues tokens at
        `https://acc.keycloak.open-regels.nl/realms/ronl/protocol/openid-connect/token`.
      flows:
        clientCredentials:
          tokenUrl: https://keycloak.open-regels.nl/realms/ronl/protocol/openid-connect/token
          scopes: {}

    mediaAggregatorKey:
      type: http
      scheme: bearer
      description: |
        A shared secret for the media aggregator, NOT a Keycloak token. Checked
        only when the server sets MEDIA_AGGREGATOR_ACCEPT_KEY; when it does not,
        /media-aggregator/search is open. Applies to that operation alone.
...
security:
  - bearerAuth: []